split up utils.tsx into smaller files
