+++ react/src/pages/utils/typings.tsx
@@ -1,0 +1,90 @@
+/*****   TYPES   *****/
+export type Counts = { [key: string]: number };
+export type KeyValue = { [key: string]: string };
+
+/*****   ENUMS   *****/
+export enum PipelineStatus {
+    PENDING = "Requested",
+    RUNNING = "Running",
+    COMPLETED = "Done",
+    ERROR = "Error",
+    CANCELLED = "Cancelled",
+}
+
+/*****   INTERFACES   *****/
+export interface Participant {
+    participant_id: string;
+    participant_codename: string;
+    family_id: string;
+    family_codename: string;
+    participant_type: string;
+    affected: boolean;
+    solved: boolean;
+    sex: string;
+    notes: string;
+    dataset_types: string[];
+    created: string;
+    created_by: number;
+    updated: string;
+    updated_by: number;
+    tissue_samples: Sample[];
+}
+export interface Sample {
+    tissue_sample_id: string;
+    extraction_date: string;
+    tissue_sample_type: string;
+    tissue_processing: string;
+    datasets: Dataset[];
+    notes: string;
+    created: string;
+    created_by: number;
+    updated: string;
+    updated_by: number;
+}
+export interface Dataset {
+    dataset_id: string;
+    participant_codename: string;
+    family_codename: string;
+    tissue_sample_type: string;
+    tissue_sample_id: string;
+    dataset_type: string;
+    input_hpf_path: string;
+    notes: string;
+    condition: string;
+    extraction_protocol: string;
+    capture_kit: string;
+    library_prep_method: string;
+    library_prep_date: string;
+    read_length: number;
+    read_type: string;
+    sequencing_id: string;
+    sequencing_centre: string;
+    batch_id: string;
+    created: string;
+    created_by: number;
+    updated: string;
+    updated_by: number;
+    discriminator: string;
+}
+export interface Analysis {
+    analysis_id: string;
+    pipeline_id: string;
+    result_hpf_path: string;
+    assignee: string;
+    requester: string;
+    analysis_state: PipelineStatus;
+    updated: string;
+    notes: string;
+    dataset_id: string;
+    qsubID: string;
+    requested: string;
+    started: string;
+    finished: string;
+    updated_by: number;
+}
+export interface Pipeline {
+    pipeline_id: number;
+    pipeline_name: string;
+    pipeline_version: string;
+    supported_types: string[];
+}
